+++ home.html
@@ -15,7 +15,6 @@
    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
    <link href="https://fonts.googleapis.com/css2?family=Lato:wght@300&display=swap" rel="stylesheet">
    <script src="./js/getBeers.js" type="module"></script>
-   <script src="./js/main.js" type="module"></script>
 </head>
 
 <body>
