+++ index.html
@@ -16,8 +16,6 @@
    <link rel="preconnect" href="https://fonts.googleapis.com">
    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
    <link href="https://fonts.googleapis.com/css2?family=Lato:wght@300&display=swap" rel="stylesheet">
-   <script src="./js/main.js" type="module"></script>
-   <!-- <script src="lodash.js"></script> -->
    <title>Home</title>
 </head>
 
@@ -52,19 +50,6 @@
                         <a class="nav-link text-dark" href="/updateBeer.html">Update Beer</a>
                      </li>
                   </ul>
-               </div>
-            </nav>
-         </div>
-         <div class="col-lg-4 col-md-6 col-sm-12 d-flex flex-row-reverse">
-            <!-- SEARCH BAR -->
-            <nav class="navbar navbar-light">
-               <div class="container-fluid">
-                  <form class="d-flex">
-                     <input class="form-control me-1" type="search" placeholder="Search" aria-label="Search">
-                     <button class="btn btn-outline-success mx-1" type="submit" id="nameBtn">Search Beer</button>
-                     <button class="btn btn-outline-success mx-1" type="submit" id="typeBtn">Search Ingredient</button>
-
-                  </form>
                </div>
             </nav>
          </div>
